Backend/scripts/image_processing/requirements.txt
functions_framework
google-cloud-vision
requests
genai 
google-cloud-secret-manager
google-generativeai